MusicModel : duration ( Int -> Double) SnapKit 이용하여 Layout 설정 중

+++ MusicPlayer/MusicPlayer/Model/MusicModel.swift
@@ -11,7 +11,7 @@ struct MusicModel: Codable {
     let singer: String?
     let album: String?
     let title: String?
-    let duration: Int?
+    let duration: Double?
     let image: String?
     let file: String?
     let lyrics: String?
